Display all markers in city during onboarding.
